+++ wasm-calc6/src/Calc/Types/Type.hs
@@ -1,20 +1,20 @@
-{-# LANGUAGE DeriveFunctor      #-}
+{-# LANGUAGE DeriveFunctor #-}
 {-# LANGUAGE DerivingStrategies #-}
-{-# LANGUAGE OverloadedStrings  #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Calc.Types.Type (Type (..), TypePrim (..)) where
 
-import           Calc.Types.TypeVar
+import Calc.Types.TypeVar
 import qualified Data.List.NonEmpty as NE
-import           GHC.Natural
-import qualified Prettyprinter      as PP
+import GHC.Natural
+import qualified Prettyprinter as PP
 
 data TypePrim = TBool | TInt | TFloat
   deriving stock (Eq, Ord, Show)
 
 instance PP.Pretty TypePrim where
-  pretty TBool  = "Boolean"
-  pretty TInt   = "Integer"
+  pretty TBool = "Boolean"
+  pretty TInt = "Integer"
   pretty TFloat = "Float"
 
 -- | resolved types
@@ -34,7 +34,8 @@ instance PP.Pretty (Type ann) where
     where
       prettyArgs = PP.concatWith (PP.surround PP.comma) (PP.pretty <$> args)
   pretty (TUnificationVar _ i) = "U" <> PP.pretty i
-  pretty (TContainer _ as) | length as == 1 =
-    "Box(" <> PP.pretty (NE.head as) <> ")"
+  pretty (TContainer _ as)
+    | length as == 1 =
+        "Box(" <> PP.pretty (NE.head as) <> ")"
   pretty (TContainer _ as) =
     "(" <> PP.cat (PP.punctuate "," (PP.pretty <$> NE.toList as)) <> ")"
