+++ wasm-calc6/test/Test/Wasm/WasmSpec.hs
@@ -2,16 +2,16 @@
 
 module Test.Wasm.WasmSpec (spec) where
 
-import           Calc.Linearity            (validateModule)
-import           Calc.Parser
-import           Calc.Typecheck
-import           Calc.Wasm.FromExpr
-import           Calc.Wasm.Run
-import           Calc.Wasm.ToWasm
-import           Data.Foldable             (traverse_)
-import           Data.Text                 (Text)
+import Calc.Linearity (validateModule)
+import Calc.Parser
+import Calc.Typecheck
+import Calc.Wasm.FromExpr
+import Calc.Wasm.Run
+import Calc.Wasm.ToWasm
+import Data.Foldable (traverse_)
+import Data.Text (Text)
 import qualified Language.Wasm.Interpreter as Wasm
-import           Test.Hspec
+import Test.Hspec
 
 testCompileExpr :: (Text, Wasm.Value) -> Spec
 testCompileExpr (input, result) = it (show input) $ do
@@ -36,8 +36,7 @@ spec :: Spec
 spec = do
   describe "WasmSpec" $ do
     let testVals =
-          [
-            ("42", Wasm.VI64 42),
+          [ ("42", Wasm.VI64 42),
             ("(1 + 1)", Wasm.VI64 2),
             ("1 + 2 + 3 + 4 + 5 + 6", Wasm.VI64 21),
             ("6 * 6", Wasm.VI64 36),
@@ -113,7 +112,7 @@ spec = do
               Wasm.VI64 42
             ),
             ("let _ = 1; 2", Wasm.VI64 2),
-            ("let Box(a) = Box(42); a", Wasm.VI64 42) ,
+            ("let Box(a) = Box(42); a", Wasm.VI64 42),
             ("let Box(a) = Box(1.23); let _ = a; 23", Wasm.VI64 23),
             ("let (a,b) = (1,2); a + b", Wasm.VI64 3),
             ("let Box(Box(a)) = Box(Box(101)); a", Wasm.VI64 101),
